Moved model step dialog to experiments new folder
